chore(refactoring): Extract ReflectionUtil into dedicated module
