+++ connector-runtime/connector-runtime-core/pom.xml
@@ -41,6 +41,11 @@
     </dependency>
 
     <dependency>
+      <groupId>io.camunda.connector</groupId>
+      <artifactId>connector-utils</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>io.camunda</groupId>
       <artifactId>camunda-client-java</artifactId>
       <version>${version.camunda}</version>
