+++ element-template-generator/core/src/main/java/io/camunda/connector/generator/java/util/TemplatePropertiesUtil.java
@@ -16,7 +16,7 @@
  */
 package io.camunda.connector.generator.java.util;
 
-import static io.camunda.connector.api.reflection.ReflectionUtil.getAllFields;
+import static io.camunda.connector.util.reflection.ReflectionUtil.getAllFields;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.camunda.connector.generator.dsl.*;
